feat: migrate supabase to single instance; initialized lazily + improved compression + supabase init on startup

+++ Illusive/src/db/supabase.ts
@@ -2,19 +2,25 @@ import { createClient, type SupportedStorage } from "@supabase/supabase-js";
 import { MMKV } from 'react-native-mmkv';
 import type { Database } from './database.types';
 
-const storage = new MMKV({ id: 'supabase-storage' })
+let supabase_instance: ReturnType<typeof createClient<Database>>;
 
-const mmkv_storage_config = {
-    setItem: (key, data) => storage.set(key, data),
-    getItem: (key) => storage.getString(key) ?? null,
-    removeItem: (key) => storage.delete(key),
-} satisfies SupportedStorage;
+export function supabase(): typeof supabase_instance{
+    if(supabase_instance) return supabase_instance;
+    const storage = new MMKV({ id: 'supabase-storage' })
 
-export const supabase = createClient<Database>(process.env.EXPO_PUBLIC_SUPABASE_PROJECT_URL ?? '', process.env.EXPO_PUBLIC_SUPABASE_PUBLIC_KEY ?? '', {
-    auth: {
-        storage: mmkv_storage_config,
-        autoRefreshToken: true,
-        persistSession: true,
-        detectSessionInUrl: false
-    }
-});
+    const mmkv_storage_config = {
+        setItem: (key, data) => storage.set(key, data),
+        getItem: (key) => storage.getString(key) ?? null,
+        removeItem: (key) => storage.delete(key),
+    } satisfies SupportedStorage;
+
+    supabase_instance = createClient<Database>(process.env.EXPO_PUBLIC_SUPABASE_PROJECT_URL ?? '', process.env.EXPO_PUBLIC_SUPABASE_PUBLIC_KEY ?? '', {
+        auth: {
+            storage: mmkv_storage_config,
+            autoRefreshToken: true,
+            persistSession: true,
+            detectSessionInUrl: false
+        }
+    });
+    return supabase_instance;
+}
